refactor: addressing review
