--- networkx/algorithms/polynomials/__init__.py
@@ -1,1 +1,0 @@
-from .tutte import *
