+++ requirements/default.txt
@@ -2,5 +2,4 @@ numpy>=1.19
 scipy>=1.8
 matplotlib>=3.4
 pandas>=1.3
-mpmath>=1.2
 sympy>=1.0
